Project nearly done.

+++ app/Http/Controllers/ProjectsController.php
@@ -5,12 +5,22 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Project;
+use Mail;
+Use App\Mail\ProjectCreated;
 
 class ProjectsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
-        $projects = Project::all();
+
+        // $projects = Project::where('owner_id', auth()->id())->get();
+
+        $projects = auth()->user()->projects;
 
         return view('projects.index', compact('projects'));
     }
@@ -22,21 +32,22 @@ class ProjectsController extends Controller
 
     public function show(Project $project)
     {
-
+        $this->authorize('view', $project);
         return view('projects.show', compact('project'));
 
     }
 
     public function edit(Project $project)
     {
-
+        $this->authorize('view', $project);
         return view('projects.edit', compact('project'));
 
     }
 
     public function update(Project $project)
     {
-
+       $project->update($this->validateProject());
+        $this->authorize('view', $project);
         $project->update(request(['title', 'description']));
 
         return redirect('/projects');
@@ -45,6 +56,8 @@ class ProjectsController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('view', $project);
+
         $project->delete();
 
         return redirect('/projects');
@@ -53,14 +66,37 @@ class ProjectsController extends Controller
 
     public function store()
     {
-        request()->validate([
-            'title' => ['required', 'min:3'],
-            'description' => 'required'
-        ]);
 
-        Project::create(request(['title','description']));
+        $attributes = ($this->validateProject());
+
+       $attributes['owner_id'] = auth()->id();
+
+       $project = Project::create($attributes);
+
+       Mail::to($project->owner->email)->send(
+           new ProjectCreated($project)
+       );
+
 
         return redirect('/projects');
+
+        /*
+       request()->validate([
+           'title' => ['required', 'min:3'],
+           'description' => 'required'
+       ]);
+
+       Project::create(request(['title','description']));
+
+       */
+    }
+
+    public function validateProject()
+    {
+        return request()->validate([
+            'title' => ['required', 'min:3'],
+            'description' => ['required', 'min:3']
+        ]);
     }
 }
